+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.4.0
 
 - Added ability to whitelist entries directly from the execution log.
+- Added Amazon Managed Workflows for Apache Airflow (:kissing_closed_eyes:) cleanup.
 
 ## 1.3.0